add delete service functionality + services context
