+++ src/components/service/service.styles.scss
@@ -112,16 +112,32 @@
         }
     }
 
-    .edit-container {
+    .actions-container {
         position: absolute;
         right: 1rem;
         top: 1rem;
+        
         background-color: $secondary-color;
         color: $primary-color;
         font-size: 3rem;
-        border-radius: 50%;
+        border-radius: 15px;
+
+        
         padding: 1rem;
         display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+
+    .edit-container {
+        color: $primary-color;
+        display: flex;
+    }
+
+    .delete-container {
+        display: flex;
+        margin-left: 1.5rem;
+        cursor: pointer;
     }
 }
