+++ src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 import './index.scss';
 import App from './App';
 import { UserProvider } from './context/user.context';
-import { ServiceProvider } from './context/service.context';
+import { ServicesProvider } from './context/services.context';
 import ScrollToTop from './utils/ScrollToTop';
 import reportWebVitals from './reportWebVitals';
 
@@ -14,9 +14,9 @@ root.render(
     <BrowserRouter>
       <ScrollToTop />
       <UserProvider>
-        <ServiceProvider>
+        <ServicesProvider>
           <App />
-        </ServiceProvider>
+        </ServicesProvider>
       </UserProvider>
     </BrowserRouter>
   </React.StrictMode>
